Added functionality to put_temp_in_symbol_table
If requested, this function will also advance the stack pointer by the
amount needed to store the temporary variable in memory.
IMPORTANT: For this to work, the STYPE parameter must have valid
is_matrix, rows and cols fields. If allocation isn't requested, the
stack_offset field must also be valid.